avisynth.cpp: copy in AvxSynth's memory checking, as a test
On a related note, it turns out that AvxSynth's support for
then-OSX and BSD was isolated to just a couple spots in the code,
this memory function and a piece in the plugin manager (which
has already been copied over and integrated as a stop-gap).

So maybe this means we get macOS and BSD support as a freebie here,
without needing to put in a lot of extra legwork.