basic Haiku support
This allows for AviSynth+ to be built natively for Haiku and
used with a corresponding build of FFmpeg (--enable-avisynth)
or avs2yuv (which needs to have its Makefile patched to allow
building on Haiku, but is otherwise fine).

Currently missing is a way of determining the executable path
so AddAutoloadDir()'s PROGRAMDIR invocation can work.

Haiku is a re-implementation of BeOS, and something like this was
brought up in a (long-since-closed) issue on Haiku's bug tracker,
way back when expecting the Ruby-based 3.0 was still a thing:
https://dev.haiku-os.org/ticket/3907